Move db pkg to server directory
