+++ backend/interfaces/api/server/model/article/linkages/to/tag/query.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/go-pg/pg"
 
-	"github.com/yoshihiro-shu/draft-backend/internal/model/article/table"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/model/article/table"
 )
 
 type (
