+++ backend/interfaces/api/server/model/article/table/type.go
